refactor: restructure source layout
- moving scripts into subdirectry
- removing obsolete scripts
- introducing BUILD_DIR var to make build and clean target consistent
- moving code into cmd resp internal

fixes issue #55
