+++ cmd/jacoco/main.go
@@ -11,7 +11,7 @@ import (
 	"time"
 
 	"github.com/hashicorp/go-retryablehttp"
-	"github.com/jenkins-x-apps/jx-app-jacoco/jacoco"
+	"github.com/jenkins-x-apps/jx-app-jacoco/internal"
 	"github.com/jenkins-x/jx/pkg/kube"
 	"k8s.io/apimachinery/pkg/fields"
 	"k8s.io/client-go/tools/cache"
