--- jacoco_analyzer_uninstall.sh
@@ -1,3 +1,0 @@
-#!/bin/sh
-NAME=jx-app-jacoco
-helm del --purge ${NAME}
