--- jacoco_maven_publish.sh
@@ -1,2 +1,0 @@
-#!/bin/sh
-jx step collect --pattern=target/site/jacoco/jacoco.xml --classifier=jacoco
