+++ scripts/curlloop.sh
