+++ scripts/jenkins-x-extension-definitions.yaml
