+++ scripts/watch.sh
